+++ src/main/java/br/edu/ifms/detran/repository/RepositoryMulta.java
@@ -7,6 +7,6 @@ import br.edu.ifms.detran.model.Multa;
 
 
 @Repository
-public interface RepositoryMulta extends JpaRepository<Multa, Long>{
+public interface RepositoryMulta extends JpaRepository<Multa, Integer>{
 
 }
